fx_pipeline/requirements.txt
requests>=2.28.0
pandas>=2.0.0
pyyaml>=6.0
pyarrow>=12.0.0
pytest>=8.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
mysql-connector-python>=8.3.0